move `log` script to the repo root
