+++ scripts/check-node-versions.mjs
@@ -3,7 +3,7 @@
 import { readFile } from 'node:fs/promises';
 import { dirname, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
-import { log, warn } from '../dotcom-rendering/scripts/env/log.js';
+import { log, warn } from './log.js';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
